Move non-entry point javascripts out of pack directory
